added grabcut demo; fixed grabCut()
git-svn-id: file://localhost/home/tolik/svnrepo/trunk/opencv@2304 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ tests/cxcore/src/arand.cpp
@@ -72,7 +72,7 @@ static double chi2_p95(int n)
     
     if( n <= 30 )
         return chi2_tab95[n-1];
-    return n + sqrt(2*n)*xp + 0.6666666666666*(xp*xp - 1);
+    return n + sqrt((double)2*n)*xp + 0.6666666666666*(xp*xp - 1);
 }
 
 bool CV_RandTest::check_pdf(const Mat& hist, double scale, double A, double B,
